added new sound for last seconds in timer
