+++ src/Components/CPR/Components/Logs.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { View } from 'react-native';
+
+/** @Componentes */
+import { BaseHorizontalLine, BaseText } from 'components';
+/** */
+
+function Logs(){
+
+  return(
+    <View>
+      <BaseHorizontalLine/>
+      <BaseText>Ultimos Eventos</BaseText>
+    </View>
+  )
+
+}
+
+export default Logs;
